requirement.txt
asgiref==3.7.2
boto3==1.34.46
botocore==1.34.46
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-crontab==0.7.1
django-import-export==3.3.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
Faker==22.6.0
jmespath==1.0.1
MarkupPy==1.14
mysqlclient==2.2.1
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
